Move `builder::curve` to `partial`
The new `partial::curve` doesn't follow the style of the `partial`
module yet. I'm going to address that in follow-up commits.
