+++ crates/fj-kernel/src/builder/mod.rs
@@ -1,6 +1,5 @@
 //! API for building objects
 
-mod curve;
 mod cycle;
 mod edge;
 mod face;
@@ -9,7 +8,6 @@ mod sketch;
 mod solid;
 
 pub use self::{
-    curve::{CurveBuilder, GlobalCurveBuilder},
     cycle::CycleBuilder,
     edge::{GlobalEdgeBuilder, HalfEdgeBuilder},
     face::FaceBuilder,
