+++ crates/fj-kernel/src/partial/curve.rs
